add api refactor
