+++ src/store/authSlice.ts
@@ -1,12 +1,13 @@
 import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
-import { api } from '../utils/api';
+import { authApi } from '../utils/api/AuthApi';
+
+const initialState = {};
 
 const signUp = createAsyncThunk(
   'auth/signUp',
   async (data: { email: string; password: string }, { rejectWithValue }) => {
     try {
-      const response = await api.signUp(data.email, data.password);
-      return response.data;
+      return authApi.signUp(data.email, data.password);
     } catch (err) {
       return rejectWithValue(err);
     }
@@ -17,8 +18,7 @@ const signIn = createAsyncThunk(
   'auth/signIn',
   async (data: { email: string; password: string }, { rejectWithValue }) => {
     try {
-      const response = await api.signIn(data.email, data.password);
-      return response;
+      return authApi.signIn(data.email, data.password);
     } catch (err) {
       return rejectWithValue(err);
     }
@@ -28,16 +28,22 @@ const signIn = createAsyncThunk(
 const forgotPassword = createAsyncThunk(
   'auth/forgotPassword',
   async (data: { email: string }) => {
-    const response = await api.forgotPassword(data.email);
-    return response.data;
+    try {
+      return authApi.forgotPassword(data.email);
+    } catch (err) {
+      return err;
+    }
   }
 );
 
 const resetPassword = createAsyncThunk(
   'auth/resetPassword',
   async (data: { token: string; password: string }) => {
-    const response = await api.resetPassword(data.token, data.password);
-    return response.data;
+    try {
+      return authApi.resetPassword(data.token, data.password);
+    } catch (err) {
+      return err;
+    }
   }
 );
 
@@ -45,8 +51,7 @@ const verifyEmail = createAsyncThunk(
   'auth/verifyEmail',
   async (data: { token: string }, { rejectWithValue }) => {
     try {
-      const response = await api.verifyEmail(data.token);
-      return response.data;
+      return authApi.verifyEmail(data.token);
     } catch (err) {
       return rejectWithValue(err);
     }
@@ -57,8 +62,7 @@ const sendVerificationCode = createAsyncThunk(
   'auth/sendVerificationCode',
   async (data, {rejectWithValue}) => {
     try {
-      const response = await api.sendVerifycationCode();
-      return response.data;
+      return authApi.sendVerifycationCode();
     } catch (err) {
       return rejectWithValue(err);
     }
@@ -69,8 +73,7 @@ const logOut = createAsyncThunk(
   'auth/logOut',
   async (data, { rejectWithValue }) => {
     try {
-      const response = await api.logOut();
-      return response.data;
+      return authApi.logOut();
     } catch (err) {
       return rejectWithValue(err);
     }
@@ -79,88 +82,8 @@ const logOut = createAsyncThunk(
 
 const authSlice = createSlice({
   name: 'auth',
-  initialState: {
-    loading: false,
-    success: false,
-  },
+  initialState,
   reducers: {},
-  extraReducers: (builder) => {
-    // Регистрация пользователя
-    builder.addCase(signUp.pending, (state) => {
-      state.loading = true;
-    });
-    builder.addCase(signUp.fulfilled, (state) => {
-      state.loading = false;
-      state.success = true;
-    });
-    builder.addCase(signUp.rejected, (state) => {
-      state.loading = false;
-    });
-
-    // Вход пользователя
-    builder.addCase(signIn.pending, (state) => {
-      state.loading = true;
-    });
-    builder.addCase(signIn.fulfilled, (state) => {
-      state.loading = false;
-      state.success = true;
-    });
-    builder.addCase(signIn.rejected, (state) => {
-      state.loading = false;
-    });
-
-    // Воставновление пароля, по email
-    builder.addCase(forgotPassword.pending, (state) => {
-      state.loading = true;
-    });
-    builder.addCase(forgotPassword.fulfilled, (state) => {
-      state.loading = false;
-      state.success = true;
-    });
-    builder.addCase(forgotPassword.rejected, (state) => {
-      state.loading = false;
-      state.success = false;
-    });
-
-    // Сохранение нового пароля, по токену
-    builder.addCase(resetPassword.pending, (state) => {
-      state.loading = true;
-    });
-    builder.addCase(resetPassword.fulfilled, (state) => {
-      state.loading = false;
-      state.success = true;
-    });
-    builder.addCase(resetPassword.rejected, (state) => {
-      state.loading = false;
-      state.success = false;
-    });
-
-    // Верификация Email по токен
-    builder.addCase(verifyEmail.pending, (state) => {
-      state.loading = true;
-    });
-    builder.addCase(verifyEmail.fulfilled, (state) => {
-      state.loading = false;
-      state.success = true;
-    });
-    builder.addCase(verifyEmail.rejected, (state) => {
-      state.loading = false;
-      state.success = false;
-    });
-
-    // Удаления токена - выход из профиля
-    builder.addCase(logOut.pending, (state) => {
-      state.loading = true;
-    });
-    builder.addCase(logOut.fulfilled, (state) => {
-      state.loading = false;
-      state.success = true;
-    });
-    builder.addCase(logOut.rejected, (state) => {
-      state.loading = false;
-      state.success = false;
-    });
-  },
 });
 
 const authSliceReducer = authSlice.reducer;
