+++ src/store/cardsSlice.ts
@@ -1,7 +1,7 @@
 import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
 import { ICard, ICardsState } from '../interfaces';
 import { generateRandomNumber } from '../utils/generateRandomNumber';
-import { api } from '../utils/api';
+import { api } from '../utils/api/Api';
 
 const initialState: ICardsState = {
   cards: [
