+++ src/store/userSlice.ts
@@ -1,18 +1,29 @@
 import {createSlice, createAsyncThunk} from '@reduxjs/toolkit';
 
-import {api} from '../utils/api';
+import { userApi } from '../utils/api/UserApi';
+import { authApi } from '../utils/api/AuthApi';
+
+const initialState = {
+  email: '',
+  firstName: '',
+  lastName: '',
+  isLogged: false,
+  isVerified: false
+};
 
 const getUser = createAsyncThunk('user/getUser', async () => {
-  const response = await api.getUser();
-  return response;
+  try {
+    return userApi.getUser();
+  } catch (err) {
+    return err;
+  }
 });
 
 const updateUser = createAsyncThunk(
   'user/updateUser',
   async (data: { email: string; firstName: string; lastName: string }, {rejectWithValue}) => {
     try {
-      const response = await api.updateUser(data.email, data.firstName, data.lastName);
-      return response;
+      return userApi.updateUser(data.email, data.firstName, data.lastName);
     } catch (err) {
       return rejectWithValue(err);
     }
@@ -23,8 +34,7 @@ const updateProfileImage = createAsyncThunk(
   'user/updateProfileImage',
   async (data : { formData: FormData }, { rejectWithValue }) => {
     try {
-      const response = await api.uploadProfileImage(data.formData);
-      return response.data;
+      return authApi.uploadProfileImage(data.formData);
     } catch (err) {
       return rejectWithValue(err);
     }
@@ -35,8 +45,7 @@ const getProfileImage = createAsyncThunk(
   'user/getProfileImage',
   async (data, { rejectWithValue }) => {
     try {
-      const response = await api.getProfileImage();
-      return response.data;
+      return authApi.getProfileImage();
     } catch (err) {
       return rejectWithValue(err);
     }
@@ -45,23 +54,13 @@ const getProfileImage = createAsyncThunk(
 
 const userSlice = createSlice({
   name: 'user',
-  initialState: {
-    email: '',
-    firstName: '',
-    lastName: '',
-    loading: false,
-    success: false,
-    isLogged: false,
-    isVerified: false
-  },
+  initialState,
   reducers: {
     signInMockUser(state, action) {
       state.email = action.payload.email;
       state.firstName = action.payload.firstName;
       state.lastName = action.payload.lastName;
       state.isLogged = true;
-      state.loading = false;
-      state.success = true;
     },
 
     updateStatus(state, action) {
@@ -70,12 +69,7 @@ const userSlice = createSlice({
   },
   extraReducers: (builder) => {
     // Получение юзера, если есть token
-    builder.addCase(getUser.pending, (state) => {
-      state.loading = true;
-    });
     builder.addCase(getUser.fulfilled, (state, action) => {
-      state.loading = false;
-      state.success = true;
       state.email = action.payload.email;
       state.firstName = action.payload.first_name;
       state.lastName = action.payload.last_name;
@@ -83,55 +77,17 @@ const userSlice = createSlice({
       state.isLogged = true;
     });
     builder.addCase(getUser.rejected, (state) => {
-      state.loading = false;
-      state.success = false;
       state.email = '';
       state.isLogged = false;
     });
 
 
-    builder.addCase(updateUser.pending, (state) => {
-      state.loading = true;
-    });
     builder.addCase(updateUser.fulfilled, (state, action) => {
-      state.loading = false;
-      state.success = true;
       state.email = action.payload.email;
       state.firstName = action.payload.first_name;
       state.lastName = action.payload.last_name;
       state.isVerified = action.payload.is_verified;
       state.isLogged = true;
-    });
-    builder.addCase(updateUser.rejected, (state) => {
-      state.loading = false;
-      state.success = false;
-    });
-
-    // Обновление фото профиля
-    builder.addCase(updateProfileImage.pending, (state) => {
-      state.loading = true;
-    });
-    builder.addCase(updateProfileImage.fulfilled, (state) => {
-      state.loading = false;
-      state.success = true;
-    });
-    builder.addCase(updateProfileImage.rejected, (state) => {
-      state.loading = false;
-      state.success = false;
-    });
-
-
-    // Получение фото профиля
-    builder.addCase(getProfileImage.pending, (state) => {
-      state.loading = true;
-    });
-    builder.addCase(getProfileImage.fulfilled, (state) => {
-      state.loading = false;
-      state.success = true;
-    });
-    builder.addCase(getProfileImage.rejected, (state) => {
-      state.loading = false;
-      state.success = false;
     });
   },
 });
